requirements.txt
playwright
beautifulsoup4
python-dotenv
requests